requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-environ==0.12.0
djangorestframework==3.16.0
environs==14.2.0
kombu==5.5.3
marshmallow==4.0.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
